Changed data type for xp
- xp data type: integer -> decimal
- changed weight: 2/min -> 1.2/min

+++ models/User.js
@@ -8,8 +8,8 @@ const User = sequelize.define('User', {
         primaryKey: true,
     },
     XP: {
-        type: DataTypes.INTEGER,
-        defaultValue: 0,   
+        type: DataTypes.DECIMAL(10, 1),
+        defaultValue: 0.0,   
     },
 });
